Refactored ExportTagStates.ts to it's own file
